Don't iterate over ids, load all the entitites at once.

+++ src/SubscriptionPurge.php
@@ -57,10 +57,10 @@ class SubscriptionPurge {
       return;
     }
 
-    // TODO: maybe it's better to use loadByProperties() ?
-    $entities = array_filter(array_map(function ($id) {
-      return $this->entityStorage->load($id);
-    }, $ids));
+    $entities = $this->entityStorage->loadMultiple($ids);
+    if (empty($entities)) {
+      return;
+    }
 
     try {
       $this->entityStorage->delete($entities);
